rename folders to "fans"
