+++ include/fans/FanCurve.h
